Add update functionality for bank accounts in server, client, and UI components

+++ src/server/CashflowNet.Server.Domain/Services/BankAccountService.cs
@@ -30,7 +30,22 @@ public class BankAccountService : IBankAccountService
     
     public async Task DeleteBankAccount(Guid id)
     {
+        var target = await _dbContext.BankAccounts.FindAsync(id);
+        if (target is null)
+            throw new Exception("Bank account not found");
+        
         _dbContext.BankAccounts.Remove(_dbContext.BankAccounts.First(x => x.Id == id));
+        
+        await _dbContext.SaveChangesAsync();
+    }
+    
+    public async Task UpdateBankAccount(UpdateBankAccountDto bankAccount)
+    {
+        var target = await _dbContext.BankAccounts.FindAsync(bankAccount.Id);
+        if (target is null)
+            throw new Exception("Bank account not found");
+        
+        target.Name = bankAccount.Name;
         
         await _dbContext.SaveChangesAsync();
     }
